+++ src/backend/api/v1/Users.ts
@@ -5,6 +5,7 @@ import * as fs from "fs";
 import {User} from "../../models/UserModel";
 import {IUser} from "../../../shared/ModelInterfaces";
 import { useEffect } from 'react';
+import {checkIsValidObjectId} from "../../shared/Middlewares";
 const multer = require('multer');
 import {applyRecommendation} from '../../shared/Helpers'
 
@@ -16,7 +17,12 @@ usersRouter.get('/', (req, res) => {
 usersRouter.get('/all', (req, res) => {
     const userList = User.find({});
     userList.exec()
-        .then((users: IUser[]) => res.send(users));
+        .then((user: IUser[]) => {
+            res.json(user);
+        })
+        .catch(() => {
+            res.status(500).json({message: `Failed to get all users from database`});
+        })
 });
 
 usersRouter.post('/login', passport.authenticate('local', { failureRedirect: '/login', failureFlash: true }), (req, res, next) => {
@@ -56,7 +62,6 @@ usersRouter.post('/register', (req, res) => {
     });
 });
 
-
 usersRouter.get('/logout', (req, res, next) => {
     req.logout();
     if (req.session) {
@@ -65,7 +70,6 @@ usersRouter.get('/logout', (req, res, next) => {
         });
     }
 });
-
 
 usersRouter.get('/:username',  (req, res, next)=> {
     if(!req.isAuthenticated()){
@@ -85,6 +89,19 @@ usersRouter.get('/:username',  (req, res, next)=> {
     } else {
         res.status(401).json({'message': 'Not Authorized'});
     }
+});
+
+usersRouter.get('/ids/:userId', checkIsValidObjectId, (req, res, next) => {
+    const userId = req.params.userId;
+    const query = User.findById(userId);
+    return query.exec()
+        .then((user: IUser | null) => {
+            res.json(user);
+        })
+        .catch((err: Error) => {
+            console.error(err);
+            res.status(500).json({message: 'Failed getting post.'})
+        });
 });
 
 // update user with username
@@ -115,12 +132,8 @@ const storage = multer.diskStorage({
 const upload = multer({ storage });
 
 usersRouter.post('/uploadAvatar', upload.single('file'), (req, res, next) => {
-    res.send("./images/" + req.file.filename);
-});
-
-usersRouter.delete('/deleteAvatar', (req, res, next) => {
     fs.unlink('./public' + req.body.oldPath.substring(1, req.body.oldPath.length), (err) => {
-        res.send('Deleted')
+        res.send("./images/" + req.file.filename);
     });
 });