+++ src/frontend/posts_page/components/HomePage.tsx
@@ -2,10 +2,14 @@ import React from "react";
 import InputBlock from "./InputBlock";
 import ComponentsContainer from "./ComponentsContainer";
 import SearchBlock from "./SearchBlock";
-import {IUser} from "./UserBlock";
+import {getPosts, getPostsByIds} from "../../shared/globleFunctions";
+import {connect} from "react-redux";
+import {loadPosts} from "../actions";
+import {IUser} from "../../../shared/ModelInterfaces";
 
 interface IHomePageProps {
-    user: IUser
+    user: IUser,
+    loadPosts: any
 }
 
 interface IHomePageState {
@@ -15,6 +19,7 @@ interface IHomePageState {
 export enum IComponentsType {
     posts = 'posts',
     users = 'users',
+    personal = 'personal'
 }
 
 class HomePage extends React.Component<IHomePageProps, IHomePageState> {
@@ -25,6 +30,11 @@ class HomePage extends React.Component<IHomePageProps, IHomePageState> {
         }
     }
 
+    async componentDidMount() {
+        let postList: any[] = await getPosts();
+        this.props.loadPosts(postList);
+    }
+
     getPosts = () => {
         this.setState({componentsType: IComponentsType.posts})
     };
@@ -33,19 +43,27 @@ class HomePage extends React.Component<IHomePageProps, IHomePageState> {
         this.setState({componentsType: IComponentsType.users})
     };
 
+    getPersonal = () => {
+        this.setState({componentsType: IComponentsType.personal})
+    };
+
     render() {
         return (
             <div id="post-blog-page">
-                <SearchBlock user={this.props.user} getPosts={this.getPosts} getUsers={this.getUsers} />
+                <SearchBlock user={this.props.user} getPosts={this.getPosts} getUsers={this.getUsers} getPersonal={this.getPersonal}/>
                 <div className="home-page-body">
                     <InputBlock user={this.props.user} />
                     <ComponentsContainer registeredUser={this.props.user} componentsType={this.state.componentsType} />
                 </div>
-
             </div>
         );
     }
 
 }
 
-export default HomePage;
+const mapStateToProps = (state: {postList: any }) => {
+    return {
+        postList: state.postList
+    };
+};
+export default connect(mapStateToProps, {loadPosts})(HomePage);