+++ src/frontend/posts_page/components/PostBlock.tsx
@@ -2,15 +2,21 @@ import React from "react";
 import {IUserProps} from "../../shared/interfaces/IUserProps";
 import Dropdown from "react-bootstrap/Dropdown";
 import {connect} from "react-redux";
-import {addLike, hidePost, undoLike} from "../actions";
+import {updateLike, deletePost, loadSavedPosts, loadHiddenPosts} from "../actions";
 import CommentsContainer from "./CommentsContainer";
 import CommentInputBar from "./CommentInputBar";
+import {IUser} from "../../../shared/ModelInterfaces";
+import {loadUserInfo} from "../../settings/actions";
+import {getPostsByIds} from "../../shared/globleFunctions";
 
 export interface IPostBlockProps {
-    post: IPost,
-    addLike: any,
-    undoLike: any,
-    hidePost: any
+    post: any,
+    updateLike: any,
+    userInfo: IUser,
+    deletePost: any,
+    loadUserInfo: any,
+    loadSavedPosts: any,
+    loadHiddenPosts: any
 }
 
 export interface IPost extends IUserProps{
@@ -21,21 +27,12 @@ export interface IPost extends IUserProps{
   avatarPath: string,
   image: string,
   numLikes: number,
-  comments: Comment[],
+  comments: any[],
   type: string,
   visibility: string,
   tags: string[],
   liked: boolean,
   hidden: boolean
-}
-
-export interface Comment {
-  time: string,
-  name: string,
-  detail: string,
-  avatar: string,
-  image: string,
-  visibility: string,
 }
 
 interface IPostBlockState {
@@ -52,21 +49,84 @@ class PostBlock extends React.Component<IPostBlockProps, IPostBlockState> {
     };
   }
 
-  markLike = (post: IPost) => {
-    if(post.liked) {
-      this.props.undoLike(post.id);
-    } else {
-      this.props.addLike(post.id);
-    }
+  markLike = async () => {
+      let update = {
+          likedUserIds: []
+      };
+      const likes = this.props.post.likedUserIds.slice();
+      if(likes.includes(this.props.userInfo._id)) {
+          likes.splice(likes.indexOf(this.props.userInfo._id), 1);
+      } else {
+          likes.push(this.props.userInfo._id);
+      }
+      update.likedUserIds = likes;
+      let response = await fetch('/api/v1/posts/' + this.props.post.id, {method: 'PATCH',
+              headers: {
+                  'Accept': 'application/json',
+                  'Content-Type': 'application/json',
+              },
+              body: JSON.stringify(update)
+          });
+      let responseData = await response.json();
+      this.props.updateLike(responseData.likedUserIds, responseData._id);
   };
 
-  hidePost = (post: IPost) => {
+  savePost = async () => {
+      let update = {
+          savedPostIds: this.props.userInfo.savedPostIds.slice()
+      };
+      if (this.props.userInfo.savedPostIds.includes(this.props.post.id)) {
+          update.savedPostIds.splice(update.savedPostIds.indexOf(this.props.post.id), 1);
+      } else {
+          update.savedPostIds.push(this.props.post.id);
+      }
+      let response = await fetch('/api/v1/users/' + this.props.userInfo.username, {
+          method: 'PATCH',
+          headers: {
+              'Accept': 'application/json',
+              'Content-Type': 'application/json',
+          },
+          body: JSON.stringify(update)
+      });
+      let responseData = await response.json();
+      this.props.loadUserInfo(responseData);
+      this.props.loadSavedPosts(await getPostsByIds(responseData.savedPostIds))
+  };
+
+  hidePost = async () => {
+      let update = {
+          hiddenPostIds: this.props.userInfo.hiddenPostIds.slice()
+      };
+      if (this.props.userInfo.hiddenPostIds.includes(this.props.post.id)) {
+          update.hiddenPostIds.splice(update.hiddenPostIds.indexOf(this.props.post.id), 1);
+      } else {
+          update.hiddenPostIds.push(this.props.post.id);
+      }
+      let response = await fetch('/api/v1/users/' + this.props.userInfo.username, {
+          method: 'PATCH',
+          headers: {
+              'Accept': 'application/json',
+              'Content-Type': 'application/json',
+          },
+          body: JSON.stringify(update)
+      });
+      let responseData = await response.json();
+      this.props.loadUserInfo(responseData);
+      this.props.loadHiddenPosts(await getPostsByIds(responseData.hiddenPostIds));
       this.setState({postHidden: !this.state.postHidden});
-      this.props.hidePost(post.id)
   };
 
   displayComment = () => {
       this.setState({showComments: !this.state.showComments});
+  };
+
+  deletePost = async () => {
+      try{
+          await fetch('/api/v1/posts/' + this.props.post.id, {method: 'DELETE'});
+          this.props.deletePost(this.props.post.id);
+      } catch (e) {
+          console.log(e);
+      }
   };
 
   render() {
@@ -75,10 +135,13 @@ class PostBlock extends React.Component<IPostBlockProps, IPostBlockState> {
       background: 'white',
       borderColor: 'white'
     };
+    const deleteButton = this.props.post.userId === this.props.userInfo._id ?
+        <Dropdown.Item className="profile-drop-down-button" onClick={this.deletePost}>
+            <span className={'glyphicon glyphicon-remove'} /> Delete Post</Dropdown.Item> : "";
     return(<div className="post-block" key={this.props.post.id}>
         <div className="hidden-post" style={this.props.post.hidden ? {display: 'block'} : {display: 'none'}}>
             <span className="hidden-post-title">Post hidden</span>
-            <button className="undo-hide-post" onClick={() => this.hidePost(this.props.post)}>Undo</button>
+            <button className="undo-hide-post" onClick={this.hidePost}>Undo</button>
         </div>
         <div style={this.props.post.hidden ? {display: 'none'} : {display: 'block'}}>
             <div className="profile-photo-block">
@@ -93,12 +156,13 @@ class PostBlock extends React.Component<IPostBlockProps, IPostBlockState> {
                             v
                         </Dropdown.Toggle>
                         <Dropdown.Menu>
-                            <Dropdown.Item className="profile-drop-down-button">
+                            <Dropdown.Item className="profile-drop-down-button" onClick={this.savePost}>
                                 <span className={'fa fa-bookmark-o'} /> Save Post</Dropdown.Item>
-                            <Dropdown.Item className="profile-drop-down-button" onClick={() => this.hidePost(this.props.post)}>
+                            <Dropdown.Item className="profile-drop-down-button" onClick={this.hidePost}>
                                 <span className={'fa fa-times-rectangle-o'} /> Hide Post</Dropdown.Item>
                             <Dropdown.Item className="profile-drop-down-button">
                                 <span className={'fa fa-exclamation-triangle'} /> Report Post</Dropdown.Item>
+                            {deleteButton}
                         </Dropdown.Menu>
                     </Dropdown>
                 </div>
@@ -109,8 +173,8 @@ class PostBlock extends React.Component<IPostBlockProps, IPostBlockState> {
                     {this.props.post.image ? <img className="inserted-image" src={this.props.post.image} alt={''}/>: ''}
                 </div>
                 <div className="interaction-buttons">
-                    <button className="like-button" onClick={() => this.markLike(this.props.post)}>
-                        <span className={'fa fa-thumbs-o-up'} /> Like {this.props.post.numLikes}</button>
+                    <button className="like-button" onClick={this.markLike}>
+                        <span className={'fa fa-thumbs-o-up'} /> Like {this.props.post.likedUserIds.length}</button>
                     <button className="comment-button" onClick={this.displayComment}>
                         <span className={'fa fa-commenting-o'} /> Comment {this.props.post.comments.length}</button>
                     <button className="share-button">
@@ -119,15 +183,18 @@ class PostBlock extends React.Component<IPostBlockProps, IPostBlockState> {
                 <div style={this.state.showComments ? {display: 'block'} : {display: 'none'}}>
                     <CommentsContainer comments={this.props.post.comments}/>
                 </div>
-                <CommentInputBar post={this.props.post} />
+                <CommentInputBar post={this.props.post} user={this.props.userInfo} />
             </div>
         </div>
     </div>)
   }
 }
 
-const mapStateToProps = (state: { postList: any}) => {
-    return {postList: state.postList};
+const mapStateToProps = (state: { postList: any, userInfo: any}) => {
+    return {
+        postList: state.postList,
+        userInfo: state.userInfo
+    };
 };
 
-export default connect(mapStateToProps, {addLike, undoLike, hidePost})(PostBlock);
+export default connect(mapStateToProps, {updateLike, deletePost, loadUserInfo, loadSavedPosts, loadHiddenPosts})(PostBlock);